pyproject.toml
[tool.poetry]
name = "django-rest-auth-ex"
version = "0.1.0"
description = ""
authors = ["hbyyy <qjaduddl94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.8"
djangorestframework = "^3.11.0"
djangorestframework-jwt = "^1.11.0"
django-rest-auth = {extras = ["with_social"], version = "^0.9.5"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
